python-requirements-mem0.txt
mem0ai==1.0.11
google-genai>=1.14,<2

